refactored file structure, added handler for public dir serving
