--- datastore/index.html
@@ -1,154 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="icon" href="img/fast-cat.png">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Dancing+Script:wght@500;600;700&family=Open+Sans&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap"
-          rel="stylesheet">
-    <link rel="stylesheet" href="../public/css/style.css?v=1.1">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-            crossorigin="anonymous"></script>
-    <title>Home</title>
-    <meta content="fastfood cataring company - fast-cat" name="description">
-    <meta content="fastfood, cataring, fast-cat" name="key-words">
-</head>
-<body>
-
-<header class="first-header">
-    <div class=""></div>
-    <div class="section-decoration section-decoration-120"
-         style="background-image: url(../public/img/section-decor/divider1_desktop.svg);">
-        <h1 class="text-center dancing go-in-rocket"><span style="color: #ff0000;">Fast</span><span style="color: #333">Cat</span>
-        </h1></div>
-</header>
-<header class="header-support"></header>
-<header class="second-header ">
-
-    <nav class="navbar navbar-expand-md navbar-light bg-yellow-custom">
-        <div class="container-fluid">
-            <a class="navbar-brand" href="/public/index.html"><img class="logo-cat" src="../public/img/fast-cat.png"
-                                                                   alt="fast cat - symbol of fast catering"></a>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                    data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
-                    aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-                    <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="/public/index.html">Home</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/public/about.htmlut.html">About</a>
-                    </li>
-                    <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button"
-                           data-bs-toggle="dropdown" aria-expanded="false">
-                            Get started
-                        </a>
-                        <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                            <li><a class="dropdown-item" href="#">Log in</a></li>
-                            <li><a class="dropdown-item" href="/public/registration.html">Registration</a></li>
-
-                        </ul>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/public/contacts.htmlts.html">Contacts</a>
-                    </li>
-
-                </ul>
-                <form class="d-flex">
-                    <input class="form-control me-2 form-control-custom-search" type="search" placeholder="Search"
-                           aria-label="Search">
-                    <button class="btn btn-outline-success custom-search" type="submit">
-                        🔎
-                    </button>
-                </form>
-
-            </div>
-            <a class="navbar-brand" href="/public/cart.htmlrt.html"><img class="cart" src="../public/img/trolley.png"
-                                                                         alt="fast cat - symbol of fast catering"></a>
-        </div>
-    </nav>
-
-
-</header>
-<header><div class="section-decoration section-decoration-50 "
-             style="background-image: url(../public/img/section-decor/header-wave2.svg);">
-
-    </div></header>
-<main>
-    <div class="main-wrapper-content">
-        <div class="container">
-            <div class="row">
-                <h1>Hello World</h1>
-                <p>{{.}}</p>
-
-            </div>
-        </div>
-    </div>
-
-</main>
-
-<footer>
-    <section class="section-decoration section-decoration-120"
-             style="background-image: url(../public/img/section-decor/footer-wave.svg)"></section>
-    <div class="wrapper-footer">
-        <div class="container">
-            <div class="row">
-
-                <div class="col-md-3">
-                    <h5>Make business with us</h5>
-                    <ul>
-                        <li><a href="#">Vacancy</a></li>
-                        <li><a href="#">Partners shops</a></li>
-                        <li><a href="#">Delivery</a></li>
-                        <li><a href="#">FastCat Business</a></li>
-                    </ul>
-                </div>
-                <div class="col-md-3">
-                    <h5>Useful reference</h5>
-                    <ul>
-                        <li><a href="#">About us</a></li>
-                        <li><a href="#">FQ</a></li>
-                        <li><a href="#">Blog</a></li>
-                        <li><a href="#">Contact us</a></li>
-                        <li><a href="#">Safety</a></li>
-                    </ul>
-                </div>
-                <div class="col-md-3">
-                    <h5>Social net</h5>
-                    <ul>
-                        <li><a href="#">Facebook</a></li>
-                        <li><a href="#">Twitter</a></li>
-                        <li><a href="#">Instagram</a></li>
-                    </ul>
-                </div>
-                <div class="col-md-3">
-                    <h5>ADDITIONAL</h5>
-                    <ul>
-                        <li><a href="#">RULES & CONDITIONS</a></li>
-                        <li><a href="#">POLICY</a></li>
-
-                    </ul>
-                </div>
-
-            </div>
-        </div>
-        <hr>
-        <div class="footer-low-writing"><span>2021 </span><span> Powered by <i><strong style="color: #fff; opacity: 1;"> Go</strong></i></span>
-        </div>
-    </div>
-</footer>
-<script src="../public/js/script.js?v=1.1"></script>
-</body>
-</html>
