+++ internal/model/product.go
