+++ public/registration.html
@@ -33,7 +33,7 @@
 
     <nav class="navbar navbar-expand-md navbar-light bg-yellow-custom">
         <div class="container-fluid">
-            <a class="navbar-brand" href="/public/index.htmlex.html"><img class="logo-cat" src="img/fast-cat.png"
+            <a class="navbar-brand" href="/index.html"><img class="logo-cat" src="img/fast-cat.png"
                                                                           alt="fast cat - symbol of fast catering"></a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                     data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
@@ -43,10 +43,10 @@
             <div class="collapse navbar-collapse" id="navbarSupportedContent">
                 <ul class="navbar-nav me-auto mb-2 mb-lg-0">
                     <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="/public/index.htmlex.html">Home</a>
+                        <a class="nav-link active" aria-current="page" href="/index.html">Home</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="/public/about.htmlut.html">About</a>
+                        <a class="nav-link" href="/about.html">About</a>
                     </li>
                     <li class="nav-item dropdown">
                         <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button"
@@ -55,12 +55,12 @@
                         </a>
                         <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
                             <li><a class="dropdown-item" href="#">Log in</a></li>
-                            <li><a class="dropdown-item" href="/public/registration.html">Registration</a></li>
+                            <li><a class="dropdown-item" href="/registration.html">Registration</a></li>
 
                         </ul>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="/public/contacts.htmlts.html">Contacts</a>
+                        <a class="nav-link" href="/contacts.html">Contacts</a>
                     </li>
 
                 </ul>
@@ -73,13 +73,11 @@
                 </form>
 
             </div>
-            <a class="navbar-brand" href="/public/cart.htmlrt.html"><img class="cart" src="img/trolley.png"
+            <a class="navbar-brand" href="/cart.html"><img class="cart" src="img/trolley.png"
                                                                          alt="fast cat - symbol of fast catering"></a>
         </div>
 
     </nav>
-
-
 </header>
 <header><div class="section-decoration section-decoration-50 "
              style="background-image: url(img/section-decor/header-wave2.svg);">
